requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
openai>=1.12.0
pinecone-client>=3.0.1
PyPDF2==3.0.1
python-docx==0.8.11
tiktoken==0.5.2